update detail admin

+++ manageApp/Form4.cs
@@ -16,6 +16,7 @@ namespace manageApp
         {
             InitializeComponent();
             toolStripStatusLabel3.Text = DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss");
+            show_Table();
         }
         private void show_Table()
         {
@@ -62,14 +63,21 @@ namespace manageApp
         private void 删除ToolStripMenuItem_Click(object sender, EventArgs e)
         {
             DialogResult result = MessageBox.Show("确定要删除吗？", "提示", MessageBoxButtons.OKCancel);
-            string id, name;
-            id = dataGridView1.SelectedCells[0].Value.ToString();
-            name = dataGridView1.SelectedCells[1].Value.ToString();
-            string sql = "delete from tb_student where id='" + id + "' and name= '" + name + "'";
-            MessageBox.Show(sql);
-            Dao dao = new Dao();
-            dao.Excute(sql);
-            show_Table();
+            if ("ok".Equals(result.ToString()))
+            {
+                string id, name;
+                id = dataGridView1.SelectedCells[0].Value.ToString();
+                name = dataGridView1.SelectedCells[1].Value.ToString();
+                string sql = "delete from tb_student where id='" + id + "' and name= '" + name + "'";
+                MessageBox.Show(sql);
+                Dao dao = new Dao();
+                dao.Excute(sql);
+                show_Table();
+            }
+            else
+            {
+                return;
+            }
         }
     }
 }
